+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - send the request id in headers and use the open handler to retreive the dataset
  - !! modified sendDataAsHeaders() to add $useOpenHandler as the first argument
+ - OpenHandler::handle() can now take any objects implementing ArrayAccess as request data
 
 2013-09-19: